Refactor out more errors to errors.py
